Moviequest_proj/Python Microservice/requirements.txt
fastapi
uvicorn[standard]
requests
scikit-learn
python-dateutil
numpy
scipy
